Use Typhoeus instead of HTTParty as HTTP library
Typhoeus is based on `libcurl` and allows us to work around issues with
invalid responses from Wordpress because it is more forgiving of
malformed HTTP responses (in particular the missing `Content-Length`
header we've been experiencing).

It's a drop in replacement for now, but will also allow us to perform
parallel requests and use persistent connections in the future, in order
to improve performance on some of the slowest parts of Workspace.