Add basic notification banner with survey message (#199)
Adds a generic notification banner (modelled on the once removed beta banner code) and a link to a new Digital Workspace survey.

* Revert "Remove beta header (#177)"

Add the header banner back now that we will have a survey again.

* Add basic notification banner with survey message

This removes the beta banner-specific code and updates it to be a
generic notification banner. It also adds a link to a new Digital
Workspace survey.