Fix notification banner not showing on homepage
This is implemented in the most awful way imaginable, but this commit
replicates the behaviour of the way it's done across the rest of the
application for the sake of consistency.

The notification banner is returned as an attribute on *every single
content item type* in Wordpress. Every controller fetches the first
item of content and chucks it into `@global_notification`, where the
notification partial in the application layout picks up on it. The
homepage wasn't doing that. Now it does.