Use separate URL for People Finder API requests
Currently the app uses the same URL for user-facing links to People
Finder as it does for API requests. People Finder is behind a CDN in
production, and API requests should not be going through the CDN.

This adds a separate environment variable `PEOPLEFINDER_API_URL` that
is used to make API requests to People Finder.