Add file expiry warning to error report form
One of the most common error reports received by the Digital Workspace
team is that the links on the page are not working. This is due to the
file links to S3 coming through from Wordpress as signed URLs expiring
after 15 minutes - if a users has a Digital Workspace page open in the
background for longer than that, all the links end up leading to a AWS
error page.

We cannot (easily) fix the underlying cause due to the way the Wordpress
media library offloading plugin works, and the myriad ways in which the
URLs come back from the Wordpress API, so this adds a message to the
error reporting form suggesting that if this is what the user is writing
to complain about, they should try refreshing their browser first.