Replace S3 URLs in WP responses with proxy URL
We will soon be deploying a proxy that sits between S3 and our users, so that
they no longer will see signed S3 URLs on pages. This will help with the
myriad expiry issues users are experiencing.

This is not a nice solution, `gsub`ing Wordpress JSON responses, but a
stopgap until we have full control of the Wordpress environment and can
generate the correct URLs "at source".