Make asset URL replacement even less greedy
We've discovered some additional forms of URL that get wrongly replaced,
escaping outside of the HTML attributes the links are contained in.

Given that no URL will contain quotes in it, this change makes sure that
any replacement will not go beyond the HTML attribute the link is
contained within.