Simplify asset proxy URL replacement
We've now had some occurrences of unsigned S3 URLs coming through in the
WP JSON response, which cause the giant ugly regular expression to fail.

This simplifies the search&replace to no longer use a regex looking for
URLs that look like signed S3 URLs, but instead just replace occurrences
of the S3 domain. This means it will catch non-signed S3 URLs (that lack
query parameters), but on the flip side means that signed URLs will now
have the signature query params at the end. As our proxy doesn't care
about these and will silently ignore them, this is a small price to pay
for simplifying this code and no longer having to maintain an awfully
complex regex.