Ensure PF API is not called with blank user ID
We have changed People Finder to require a user ID being passed into the
people API, and this has unearthed that this app very frequently calls
it without one.

We need to figure out why this is happening, but until then, this stops
the API being called if there is no user ID to call it with.